Update middleware pckg
