+++ go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/golang/mock v1.6.0
 	github.com/iancoleman/strcase v0.2.0
 	github.com/labstack/echo v3.3.10+incompatible
+	github.com/labstack/echo/v4 v4.2.0
 	github.com/stretchr/testify v1.8.1
 	google.golang.org/grpc v1.51.0
 	gopkg.in/DataDog/dd-trace-go.v1 v1.43.1
