+++ gomockhandler.json
@@ -2,7 +2,7 @@
 	"mocks": {
 		"DatadogMetricsClient": {
 			"checksum": "NydKUvnQJ3LPAfMgNhUA9w==",
-			"source_checksum": "zu6GOBn/tWEABrE4PJxmFg==",
+			"source_checksum": "NmI2tghjdQQKsK5XAvdcSg==",
 			"mode": "SOURCE_MODE",
 			"source_mode_runner": {
 				"source": "metric/datadog.go",
@@ -19,12 +19,12 @@
 			}
 		},
 		"EchoContext": {
-			"checksum": "PSG18wg++awZ1qw/7oljqQ==",
-			"source_checksum": "pWnvl6vrYLS6YojmsOscmQ==",
+			"checksum": "gwqpPqMa/TonFyB94kC0Eg==",
+			"source_checksum": "6nT2TO09m2nsGZRXGep5Tw==",
 			"mode": "SOURCE_MODE",
 			"source_mode_runner": {
-				"source": "vendor/github.com/labstack/echo/context.go",
-				"destination": "internal/generated/mocks/labstack/echo/context.go"
+				"source": "vendor/github.com/labstack/echo/v4/context.go",
+				"destination": "internal/generated/mocks/labstack/echo/v4/context.go"
 			}
 		}
 	}
