+++ middleware/grpc/grpc_test.go
@@ -1,4 +1,4 @@
-package middleware
+package grpc
 
 import (
     "context"
