--- middleware/middleware.go
@@ -1,6 +1,0 @@
-package middleware
-
-const (
-    ddGRPCResourceName = "grpc.request"
-    ddHTTPResourceName = "http.request"
-)
